refactor(users): split 'superdesk-users' module.
`superdesk-users` becomes:
```
.
├── activity-widget
│   ├── activity.js
│   ├── configuration.html
│   ├── thumbnail.png
│   ├── thumbnail.svg
│   ├── widget-activity.html
│   └── widget-activity.less
├── controllers
│   ├── ChangeAvatarController.js
│   ├── SessionsDeleteCommand.js
│   ├── UserDeleteCommand.js
│   ├── UserEditController.js
│   ├── UserEnableCommand.js
│   ├── UserListController.js
│   ├── UserResolver.js
│   └── index.js
├── directives
│   ├── ActivityDirective.js
│   ├── ChangePasswordDirective.js
│   ├── InfoItemDirective.js
│   ├── PasswordConfirmDirective.js
│   ├── ResetPasswordDirective.js
│   ├── RolesPrivilegesDirective.js
│   ├── UserActivityDirective.js
│   ├── UserDetailsPaneDirective.js
│   ├── UserEditDirective.js
│   ├── UserInfoDirective.js
│   ├── UserListDirective.js
│   ├── UserListItemDirective.js
│   ├── UserMentioDirective.js
│   ├── UserPreferencesDirective.js
│   ├── UserPrivilegesDirective.js
│   ├── UserRolesDirective.js
│   ├── UserUniqueDirective.js
│   ├── ValidErrorDirective.js
│   ├── ValidInfoDirective.js
│   └── index.js
├── import
│   ├── tests
│   ├── views
│   └── import.js
├── services
│   ├── ProfileService.js
│   ├── UserListService.js
│   ├── UserPopupService.js
│   ├── UsersService.js
│   └── index.js
├── styles
│   ├── change-avatar.less
│   ├── settings.less
│   └── users.less
├── tests
│   ├── directives.spec.js
│   ├── profile.spec.js
│   ├── sdUserPreferences.spec.js
│   ├── sdUserPrivileges.spec.js
│   └── users.spec.js
├── views
│   ├── activity-feed.html
│   ├── activity-list.html
│   ├── change-avatar.html
│   ├── edit-form.html
│   ├── edit.html
│   ├── list.html
│   ├── mentions-menu.html
│   ├── mentions.html
│   ├── settings-privileges.html
│   ├── settings-roles.html
│   ├── settings.html
│   ├── user-details-pane.html
│   ├── user-list-item.html
│   ├── user-preferences.html
│   └── user-privileges.html
├── config.js
├── helpers.js
└── index.js

10 directories, 66 files
```
