+++ scripts/superdesk-users/activity-widget/activity.js
