+++ scripts/superdesk-users/activity-widget/configuration.html
