+++ scripts/superdesk-users/activity-widget/thumbnail.png
