+++ scripts/superdesk-users/activity-widget/widget-activity.less
