+++ scripts/superdesk-users/controllers/UserResolver.js
@@ -1,0 +1,15 @@
+/**
+ * Resolve a user by route id and redirect to /users if such user does not exist
+ */
+UserResolver.$inject = ['api', '$route', 'notify', 'gettext', '$location'];
+export function UserResolver(api, $route, notify, gettext, $location) {
+    return api.users.getById($route.current.params._id)
+        .then(null, function(response) {
+            if (response.status === 404) {
+                $location.path('/users/');
+                notify.error(gettext('User was not found, sorry.'), 5000);
+            }
+
+            return response;
+        });
+}
