+++ scripts/superdesk-users/controllers/index.js
@@ -1,0 +1,11 @@
+export { UserEditController } from './UserEditController';
+export { UserListController } from './UserListController';
+export { ChangeAvatarController } from './ChangeAvatarController';
+export { UserEnableCommand } from './UserEnableCommand';
+export { UserDeleteCommand } from './UserEnableCommand';
+export { SessionsDeleteCommand } from './SessionsDeleteCommand';
+export { UserResolver } from './UserResolver';
+
+// TODO: is this needed?
+UserRolesController.$inject = ['$scope'];
+export function UserRolesController($scope) {}
