+++ scripts/superdesk-users/directives/ActivityDirective.js
@@ -1,0 +1,6 @@
+ActivityDirective.$inject = ['asset'];
+export function ActivityDirective(asset) {
+    return {
+        templateUrl: asset.templateUrl('superdesk-users/views/activity-list.html')
+    };
+}
