+++ scripts/superdesk-users/directives/UserInfoDirective.js
@@ -1,0 +1,13 @@
+UserInfoDirective.$inject = ['userPopup'];
+export function UserInfoDirective(userPopup) {
+    return {
+        link: function(scope, element, attrs) {
+            element.addClass('user-link');
+            element.hover(function() {
+                userPopup.set(attrs.user, element, scope);
+            }, function() {
+                userPopup.close();
+            });
+        }
+    };
+}
