+++ scripts/superdesk-users/directives/UserListItemDirective.js
@@ -1,0 +1,6 @@
+UserListItemDirective.$inject = ['asset'];
+export function UserListItemDirective(asset) {
+    return {
+        templateUrl: asset.templateUrl('superdesk-users/views/user-list-item.html')
+    };
+}
