+++ scripts/superdesk-users/directives/ValidErrorDirective.js
@@ -1,0 +1,7 @@
+export function ValidErrorDirective() {
+    return {
+        link: function (scope, element) {
+            element.addClass('validation-error');
+        }
+    };
+}
