+++ scripts/superdesk-users/directives/ValidInfoDirective.js
@@ -1,0 +1,7 @@
+export function ValidInfoDirective() {
+    return {
+        link: function (scope, element) {
+            element.addClass('validation-info');
+        }
+    };
+}
