+++ scripts/superdesk-users/directives/index.js
@@ -1,0 +1,19 @@
+export { UserRolesDirective } from './UserRolesDirective';
+export { RolesPrivilegesDirective } from './RolesPrivilegesDirective';
+export { InfoItemDirective } from './InfoItemDirective';
+export { ValidErrorDirective } from './ValidErrorDirective';
+export { ValidInfoDirective } from './ValidInfoDirective';
+export { UserDetailsPaneDirective } from './UserDetailsPaneDirective';
+export { UserEditDirective } from './UserEditDirective';
+export { UserPreferencesDirective } from './UserPreferencesDirective';
+export { UserPrivilegesDirective } from './UserPrivilegesDirective';
+export { ChangePasswordDirective } from './ChangePasswordDirective';
+export { ResetPasswordDirective } from './ResetPasswordDirective';
+export { UserUniqueDirective } from './UserUniqueDirective';
+export { PasswordConfirmDirective } from './PasswordConfirmDirective';
+export { UserListDirective } from './UserListDirective';
+export { UserListItemDirective } from './UserListItemDirective';
+export { ActivityDirective } from './ActivityDirective';
+export { UserMentioDirective } from './UserMentioDirective';
+export { UserInfoDirective } from './UserInfoDirective';
+export { UserActivityDirective } from './UserActivityDirective';
