+++ scripts/superdesk-users/services/index.js
@@ -1,0 +1,4 @@
+export { UsersService } from './UsersService';
+export { UserListService } from './UserListService';
+export { UserPopupService } from './UserPopupService';
+export { ProfileService } from './ProfileService';
