+++ scripts/superdesk-users/tests/directives.spec.js
