+++ scripts/superdesk-users/tests/profile.spec.js
